cephadm: change mount point for ceph-nvmeof.conf
In the downstream nvmeof container, the nvmeof code
is rooted at /remote-source/ceph-nvmeof/app/ rather
than /src/ as it is in the upstream. This change adjusts
the mount point in the container for the ceph-nvmeof.conf
so that it matches that difference.

Signed-off-by: Adam King <adking@redhat.com>

Resolves: rhbz#2240588
(cherry picked from commit c01146d9bf15b3ea1d5789876e62ad51e9906cb4)