mgr/smb: add a cluster resource field to manage clustering
Add a new `clustering` field to the smb cluster resource. This field can
be used to select either automatic clustering with ctdb, or disable it,
or require it. The default is automatic and is based on the count value
in the placement spec. A count of 1 disables clustering and any other
value it is enabled.

Resolves: rhbz#2307978
Signed-off-by: John Mulligan <jmulligan@redhat.com>
(cherry picked from commit 4ef7a48e55910830725e1bf7b7fcbeda5e89b74d)