mgr/cephadm: make nvme-gw adds be able to handle multiple services/groups
Before this was grabbing the service spec for the first daemon
description in the list. This meant every daemon would be added
with the pool/group of whatever that spec happened to specify.
This patch grabs the spec, and therefore also the pool/group
individually for each nvmeof daemon

Signed-off-by: Adam King <adking@redhat.com>
(cherry picked from commit 2a6b105b8ae9cb63702d83461560e65b7fabd085)