mgr/vol: fix subvolume removal with retained snapshots when osd is full
The order of operation done for subvolume removal with retain-snapshot
option set to true, is reversed. The metadata is updated first
followed by a rename operation on the uuid directory. If the metadata
update operation fails, then the remove operations is failed thereby,
keeping the subvolume metadata consistent with the uuid path.

Fixes: https://tracker.ceph.com/issues/67330
Signed-off-by: Neeraj Pratap Singh <neesingh@redhat.com>