mgr/vol: separate reusable code for printing progress bars
A good amount of code from class CloneProgressReporter is generic to
getting, processing and then reporting stats through progress bar. This
generic helper code can be used not just for reporting statistics of the
async clone jobs but any async jobs. Therefore, move this code to a new
class so that it can be re-used through class inheritance. Class
PurgeProgressbar that will be introduced in the following commits will
be re-using all this code.

Signed-off-by: Rishabh Dave <ridave@redhat.com>