rgw/storage class: don't store/report STANDARD storage class.
While 'STANDARD' is a valid storage class, it is not supposed
to ever be returned when fetching an object.  This change suppresses
storing 'STANDARD' as the attribute value, so that objects
explicitly created with 'STANDARD' will in fact be indistinguishable
from those where it was implicitly set.

Fixes: https://tracker.ceph.com/issues/67786

Signed-off-by: Marcus Watts <mwatts@redhat.com>
(cherry picked from commit 120b81348c608cf70aa63e2c4d6eb130c7224e6a)

Resolves: rhbz#2300284