rgw/multisite: avoid redundant error repo entry logging
in bucket incremental sync, if a generation other than the current generation is requested,
we mark it with -EAGAIN and retry in error repo in RGWDataSyncSingleEntry(). within this block,
we check if that requested generation is lesser than the current gen, write it to error repo,
set -EAGAIN and write it to error repo in the outer function once again. don't duplicate error
repo entry addition for this condition.

Signed-off-by: Shilpa Jagannath <smanjara@redhat.com>
(cherry picked from commit a191f9a6d2997f655f580e16731615e77350a03d)
resolves rhbz#2315844