cephadm: do not trigger rollback in bootstrap unit tests
Add a context manager that wraps the normal mock ctx context manager -
this one disables bootstrap rollback by default. Use it in the tests
that call command_bootstrap. The unit tests are older than the boostrap
rollback behavior and therefore don't need it. However, this was not
previously a problem. In python 3.12 (or 3.11?) changes to the shutil module
prevent shutil.rmtree from working properly with the fake file system
that cephadm tests rely upon heavily.

Part of an effort to get ceph tox environments passing on Python 3.12.

Signed-off-by: John Mulligan <jmulligan@redhat.com>