osd/scrub: fix 'schedule-deepscrub' test asok command
The existing implementation of the 'schedule-deepscrub' Asok
command uses the set_last_deep_scrub_stamp() method to "fake"
the last-deep-scrub stamp. Unfortunately, this method also
updates the last-scrub stamp (as required for non-test usage).

Commit 9f3e18b fixed the comparator used when sorting the
scrub targets. An unintended side effect is that
following 'schedule-deepscrub' - the shallow target
is the one to be scrubbed next, instead of the deep target.

Signed-off-by: Ronen Friedman <rfriedma@redhat.com>