osd/PeeringState: rename "cancel_backfill" to "suspend_backfill"
PerringState events the are leading to `cancel_backfill()` are:
* DeferBackfill - Called if local recovery reservation is revoked
  before it completes (See AsyncResever::request_reservation on_preempt)
* UnfoundBackfill
* RemoteReservationRevokedTooFull
* RemoteReservationRevoked

In each event, we merely suspend the the backfill.
The primary will *continue* to keep trying to start this backfill
as long as the up set for the current interval includes the osds
that needs to be backfilled.

Eventually, the backfill will either succeed and complete or will
be made irrelevant due to an interval change (and essentially truly
"cancelled")

Signed-off-by: Xuehan Xu <xuxuehan@qianxin.com>