qa/cephfs: add test for concurrent readdir and rename from multiple clients
There are some reports of a possible bug in the MDS and/or the kernel
driver, which is trigerred via concurrent readdir and rename operations
on a directory from multiple clients resulting in the rename operation
getting blocked due to the kernel driver not responding to capability
revoke operation by the MDS (for the Fs cap). It is currently not known
as in which kernel versions are effected, but the reproducer involved
a LTS kernel (5.14) with latest Quincy (17.2.x).

qa/cephfs does not test with LTS kernels, so till that infrastructure
is built and available for running tests, introduce a test involving
concurrent readdir and rename operations from multiple clients running
testing kernel.

Partially-Fixes: http://tracker.ceph.com/issues/69440
Signed-off-by: Venky Shankar <vshankar@redhat.com>