interval_set: Enhancements to improve efficiency of insert() and to allow more flexibility.
The old insert was restrictive in ranges that could be added in. The new interface allows for a range to be added, whether it extends or joins other intervals.

Also change a number of interfaces to use the new insert.

Signed-off-by: Alex Ainscow <aainscow@uk.ibm.com>
(cherry picked from commit c90e94b70c69f14bcf02fe41b7ba1969b38c7da8)