mgr/cephadm: allow users to set custom max coredump size for cluster hosts
If the mgr/cephadm/set_coredump_overrides setting is set to True cephadm
will go to each host in the cluster and setup a systemd drop-in under
/etc/systemd/coredump.conf.d that sets coredumpctl parameters. Currently
the only setting is the max coredump size controlled by the
mgr/cephadm/coredump_max_size setting. If that value is modified, cephadm
will go back and rewrite this file with the new value. If new settings
are added in the future, we will also rewrite the drop-in when they change.

Signed-off-by: Adam King <adking@redhat.com>
(cherry picked from commit 7c86b4f0214394e8397e97000a3945eadaebe2b3)

Resolves: rhbz#2338598