mgr/dashboard: Multi-site replication wizard breaks when a default realm is already present in the secondary cluster.
When a default realm is already present in a secondary cluster, the multi-site automation wizard breaks and shows the progress as 0% again. This happens because the last step of the wizard is to confirm the presence of replication user created as a part of the wizard to be present in the secondary cluster and then do the further processing. We fetch the user list from the secondary cluster using get_user_list method which takes the zoneName as input param. In case there is a default realm in the secondary cluster, we need this method to take the realm name as input param as well.

Fixes: https://tracker.ceph.com/issues/69670

Resolves: rhbz#2320952

Signed-off-by: Aashish Sharma <aasharma@redhat.com>
(cherry picked from commit c96f8a77e9913a916f9cc1e9d8245990c6515402)