cephadm: rework with_cephadm_ctx fixture to use exitstack
Rework the `with_cephadm_ctx` function in fixtures.py to make use of
contextlib.ExitStack. The use of ExitStack, to quote the python docs,
"is designed to make it easy to programmatically combine other context
managers." This function makes use of many mock.patch context managers
and does already try to programmatically combine them - and the current
method is awkward.
A future change will do even more to control what gets mocked out and
so a this clean up makes things easier to read now and easier to update
later.

Signed-off-by: John Mulligan <jmulligan@redhat.com>