mgr/vol: print proper message when subvolume metadata filename is too...
long.

When combination of subvolume group name and subvolume name is longer
than 248 characters, it leads to a failure because the metadata file is
a combination of both of these along with ":", "_" and ".meta".

Currently, when this comibnation longer than 248 characters, a bunch
stacktraces are printed along with multiple errors. This confuses the
user as well as looks bad.

Fixes: https://tracker.ceph.com/issues/69865
Signed-off-by: Rishabh Dave <ridave@redhat.com>
(cherry picked from commit 82fc1e7cac3d9157d9aaf95c45ded04c1b278183)