qa/suites/upgrade/telemetry-upgrade: improve ignorelist
In this commit, I added some pattern matching for
warnings that show up in the cluster log detail that
are related to degraded PGs. In these tests, we are intentionally
restarting OSDs to upgrade them, which leads to these states
showing up in the cluster log. So, the warnings are intended
and can be ignored in the context of an upgrade.

Fixes: https://tracker.ceph.com/issues/67881
Signed-off-by: Laura Flores <lflores@ibm.com>