rbd-mirror: do not move the images to trash while the disabling is in progress
Images cannot be moved to trash if the state is disabling because its a
transient state where some of the images might have got the oportunity to
disable and some of them part of the group might still be enabled
waiting for the oportunity while a group disable is in progress.

So we wait until the state DISABLING moves to next state, and see if there are
any stale image to move into a trash queue later.

Signed-off-by: Prasanna Kumar Kalever <prasanna.kalever@redhat.com>

Resolves: rhbz#2089305