LogMonitor: set no_reply for forward MLog commands
On streach mod clusters we can see slow ops when
removing and adding osds with --zap --force when osds
connected to peon monitor and forwarding the MLog to leader.
the no_reply is set only when we are connected to the leader,
this fix will add also the other option - so no_reply set anyway.

Fixes: https://tracker.ceph.com/issues/54489
Signed-off-by: Nitzan Mordechai <nmordech@redhat.com>