mds/rename: Handle referent remote linkmerge case
For a file having hardlinks, when the primary file is
removed first before the secondary, stray reintegration
makes any secondary link available in the cache to be
primary. This is referred as linkmerge.

When the linkmerge happens, the referent inode of the
secondary needs to be removed otherwise it causes the
referent inode object leak in the data pool. This PR
handles the linkmerge scenario and removed the referent
inode of the secondary during the linkmerge.

Also, this patch does the following.
  - removes the corresponding referent inode from the
    referent_inodes list of primary/real inode
  - adds is_referent_remote() check in required places
    for linkmerge in the rename path.

Fixes: https://tracker.ceph.com/issues/54205
Signed-off-by: Kotresh HR <khiremat@redhat.com>