qa/test_backtrace: Validate remote_inode xattr is stored
Validate 'remote_inode' xattr is stored on the
referent inode upon hardlink creation if the
referent inodes feature is enabled. The value
of the xattr would be the inode number of the
primary inode.

Also validate that the referent inode is not
created on hardlink creation if the feature
is disabled.

Fixes: https://tracker.ceph.com/issues/69338
Signed-off-by: Kotresh HR <khiremat@redhat.com>