osd: Rollback changes
EC Pools create an ObjectModDesc entry in each log entry that describes
how to undo the modifcation. During peering if it is determined that
only some of the shards completed an update then these shards are
instructed to rollback the change. This ensures that each update
is either applied to all or non of the shards.i

For EC optimized pools rollback becomes a bit more complicated.
Firstly because not all shards may have been updated the rollback
needs to be more selective in what is undone. Secondly optimized
pools do not pad objects to be a multiple of the stripe width
which means shards can be different sizes. There is a single
ObjectModDesc entry that contains a set of operations applied by
every shard, individial operations include enough information
to work out what has to be done to each shard.

Signed-off-by: Bill Scales <156200352+bill-scales@users.noreply.github.com>