osd: Backfill changes - WIP crimson
Changes to backfill to use the extra information in the
object_info_t structure to work out what objects require
backfill for an EC optimization pool where sometimes
shards do not recieve updates to the OI attribute.

The object_info_t structure on the primary contains a
map of version numbers for all the other shards. This
allows the primary to compare its recorded version
number for a shard with the object_info_t version from
that shard to work out if backfill is required.

The replica shards and the messages between the replica
and the primary do not change, however the primary needs
to store more data when there are multiple backfill targets
because each target can have a different version number for
each object.

WIP: This commit doeesn't complie for Crimson yet

Signed-off-by: Bill Scales <156200352+bill-scales@users.noreply.github.com>