crimson/osd: Add scrub stubs for crimson and classic, ready for new EC
The new optimised EC code is not backward compatible withold  EC Code.
Before this commit there is some stub code which assumes that an hinfo
xattr will exist and can be used for scrub. This is no longer the case in new EC.

We plan to first make the scrubbing changes for new EC in classic and will
subsequently port to crimson. It will not look like the code here, so there is
little point in keeping it.

Additionally, add some stubs for scrub in classic optimized EC.

There will be a later PR specifically for dealing with scrubbing in
new EC which fix all the fix mes in class,

The crimson code will be fixed up at a later date and will only
support optimised EC.

Signed-off-by: Alex Ainscow <aainscow@uk.ibm.com>