DNM: mgr/vol: create dummy linkt for .meta file in v3
Keep .meta symlink in v3 when current incarnation is deleted but
subvolume dir is still exists because it was deleted with
--retain-snapshots option. Keeping helps the code for subvolume
discovery (in subvolume_base.py) to detect that it is a v3 subvolume and
not some older subvolume version. .meta symlink is kept as a broken
symlink if there is not current incarnation of subvolume yet.

Signed-off-by: Rishabh Dave <ridave@redhat.com>