rgw/d4n: this is a temp fix to get instance value in get_obj_attrs() from backend store, using read_op->prepare() method.
target_obj will be removed from get_obj_attrs
as part of this PR: https://github.com/ceph/ceph/pull/62787
and it was being used to get the object's instance
value. so using read_op->prepare(), till this gets fixed

Samarah identified that returning ENOENT when target_obj
is NULL, is incorrect and suggested removing that.

Signed-off-by: Samarah <samarah.uriarte@ibm.com>
Signed-off-by: Pritha Srivastava <prsrivas@redhat.com>