dashboard: ensure nodeenv downloaded content is owned by current user
When testing ceph builds in a container we discovered that certain files
could not be deleted by jenkins after a build. This was due to the way
the container maps IDs - files owned by the root user in the container
become owned by the "real" user/jenkins user on the "host".
However, the node tarball that is fetched and unpacked by nodeenv has
a different owner name/uid that is preserved in the tree and this id
gets mapped to something that can be managed by the "fake root" of the
container but not by the "regular" user outside the container.

The simplest workaround I can think of is to chown the tree back
to the current user and avoid leaving files on disk with uncleanly
mapped uids.

Conflicts:
	src/pybind/mgr/dashboard/frontend/CMakeLists.txt

Signed-off-by: John Mulligan <jmulligan@redhat.com>
(cherry picked from commit 24dbfb5da4813c6588f9cd199b9f527bb67f1e88)
(cherry picked from commit f7e2a85031d7e7569213a558a05189a29f04fca6)