osd: Check OI version against reference in scrub if not expected to exactly match.
With optimised EC, some shards do not receive every IO updates (partial writes).

In such cases, the OI on these "nonprimary" shards is expected to be out of date.
This commit allows for the mismatch and instead checks that the OI can be decoded
and checks the version against the shard versions in the authoritative OI.

Signed-off-by: Alex Ainscow <aainscow@uk.ibm.com>