osd: EC optimizations clean_temps too aggressive at keeping pg_temp
For optimized EC pools clean_temps was refusing to clear pg_temp if
it would result in a change of primary. This is too aggressive, it
needs to prevent a change of primary to a non-primary shard
(which will be why pg_temp was set to force select an appropriate
primary) but still needs to allow a change of primary to a
shard that is capable of becoming the primary.

Blocking changes of primary can stop PGs from merging because
the merge can't happen until both PGs have the same primary.

Signed-off-by: Bill Scales <bill_scales@uk.ibm.com>