osd: Do not apply log entry if shard not written.
This was a failed test, where the primary concluded that all objects were present
despite one missing object on the non primary shard.

The problem was caused because the log entries are sent to the unwritten shards if that
shard is missing in order to update the version number in the missing object.  However,
the log entry should not actually be added to the log.

Signed-off-by: Alex Ainscow <aainscow@uk.ibm.com>