rgw: pass list_parts_each_t function by lvalue reference
list_parts_each_t is an alias of
`fu2::unique_function<int(const Part&) const>`, which is a non copyable
function. so in theory, we cannot copy it. and in the recent version of
function2, unique_function is not coyable anymore. if we bump up the
vendored function2.hpp, the build breaks.

so, in this change, we change the virtual function of
`Object::list_parts()` from passing the plain value of
`list_parts_each_t` to rvalue reference `list_parts_each_t` so that we
don't need to copy this non-copyable function. this allows us to
keep in sync with upstream function2, and to be symantically correct
regarding to the unique-ness of the functor.

Resolves: ISCE-2472

Signed-off-by: Kefu Chai <tchaikov@gmail.com>
(cherry picked from commit 917a2dde418d519dfd8105177725ccedab3e82fe)